Name Changes and GUI Controller Update - GUI controller now inherits from controller - GUI controller named better - Minor CLI Syntax updates - main now imports according to renamed files
